third_party/WebKit/LayoutTests/fast/canvas/canvas-createImageBitmap-video-resize.html

<!DOCTYPE HTML>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
function createNewCanvas(width, height)
{
    var canvas = document.createElement("canvas");
    canvas.width = width;
    canvas.height = height;
    var ctx = canvas.getContext("2d");
    ctx.clearRect(0, 0, width, height);
    return ctx;
}

function compareBitmaps(bitmap1, bitmap2)
{
    var canvas1 = document.createElement("canvas");
    var canvas2 = document.createElement("canvas");
    canvas1.width = 50;
    canvas1.height = 50;
    canvas2.width = 50;
    canvas2.height = 50;
    var ctx1 = createNewCanvas(50, 50);
    var ctx2 = createNewCanvas(50, 50);
    ctx1.drawImage(bitmap1, 0, 0);
    ctx2.drawImage(bitmap2, 0, 0);
    var data1 = ctx1.getImageData(0, 0, 50, 50).data;
    var data2 = ctx2.getImageData(0, 0, 50, 50).data;
    var dataMatched = true;
    for (var i = 0; i < data1.length; i++) {
        if (data1[i] != data2[i]) {
            dataMatched = false;
            break;
        }
    }
    assert_false(dataMatched);
}

function checkLowResult(imageBitmap, video, sx, sy, sw, sh)
{
    var ctx1 = createNewCanvas(50, 50);
    var ctx2 = createNewCanvas(50, 50);
    ctx1.drawImage(imageBitmap, 0, 0);
    ctx2.drawImage(video, sx, sy, sw, sh, 0, 0, 40, 30);
    var data1 = ctx1.getImageData(0, 0, 50, 50).data;
    var data2 = ctx2.getImageData(0, 0, 50, 50).data;
    var dataMatched = true;
    for (var i = 0; i < data1.length; i++) {
        if (data1[i] != data2[i]) {
            dataMatched = false;
            break;
        }
    }
    assert_true(dataMatched);
}

function testImageBitmap(source)
{
    return Promise.all([
        createImageBitmap(source, {resizeWidth: 40, resizeHeight: 30, resizeQuality: "high"}),
        createImageBitmap(source, {resizeWidth: 40, resizeHeight: 30, resizeQuality: "medium"}),
        createImageBitmap(source, {resizeWidth: 40, resizeHeight: 30, resizeQuality: "low"}),
        createImageBitmap(source, {resizeWidth: 40, resizeHeight: 30, resizeQuality: "pixelated"}),
        createImageBitmap(source, 50, 50, 160, 120, {resizeWidth: 40, resizeHeight: 30, resizeQuality: "high"}),
        createImageBitmap(source, 50, 50, 160, 120, {resizeWidth: 40, resizeHeight: 30, resizeQuality: "medium"}),
        createImageBitmap(source, 50, 50, 160, 120, {resizeWidth: 40, resizeHeight: 30, resizeQuality: "low"}),
        createImageBitmap(source, 50, 50, 160, 120, {resizeWidth: 40, resizeHeight: 30, resizeQuality: "pixelated"}),
    ]).then(([noCropHigh, noCropMedium, noCropLow, noCropPixelated, cropHigh, cropMedium, cropLow, cropPixelated]) => {
        checkLowResult(noCropLow, source, 0, 0, source.videoWidth, source.videoHeight);
        checkLowResult(cropLow, source, 50, 50, 160, 120);
        // Brute-force comparison among all bitmaps is too expensive
        compareBitmaps(noCropHigh, noCropMedium);
        compareBitmaps(noCropLow, noCropPixelated);
        compareBitmaps(cropHigh, cropMedium);
        compareBitmaps(cropLow, cropPixelated);
    });
}

// HTMLVideoElement
promise_test(function() {
    var video = document.createElement("video");
    video.oncanplaythrough = function() {
        return testImageBitmap(video);
    };
    video.src = "../../compositing/resources/video.ogv";
}, 'createImageBitmap from a HTMLVideoElement with resize option.');
</script>
